+++ packages/components/src/internal/util/Date.spec.ts
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { initUnitTests } from '../enzymeTestHelpers';
+import { initUnitTests } from '../test/testHelpers';
 
 import { QueryColumn } from '../../public/QueryColumn';
 import { DATE_TYPE, DATETIME_TYPE } from '../components/domainproperties/PropDescType';
