+++ packages/components/src/public/QueryModel/withQueryModels.spec.tsx
@@ -3,8 +3,7 @@ import { mount } from 'enzyme';
 import { createMemoryHistory, InjectedRouter, Route, Router } from 'react-router';
 import { Filter } from '@labkey/api';
 
-import { sleep } from '../../internal/test/testHelpers';
-import { initUnitTests, makeQueryInfo, makeTestData } from '../../internal/enzymeTestHelpers';
+import { initUnitTests, makeQueryInfo, makeTestData, sleep } from '../../internal/test/testHelpers';
 import { MockQueryModelLoader } from '../../test/MockQueryModelLoader';
 import mixturesQueryInfo from '../../test/data/mixtures-getQueryDetails.json';
 import mixturesQuery from '../../test/data/mixtures-getQueryPaging.json';
